Update options interface
