+++ packages/electron/src/options.ts
@@ -1,0 +1,3 @@
+import { BuildOptions } from '@kano/kit-app-shell-core/lib/options';
+
+export type ElectronBuildOptions = BuildOptions;
