+++ packages/kano/src/options.ts
@@ -1,0 +1,5 @@
+import { ElectronBuildOptions } from '@kano/kit-app-shell-electron/lib/options';
+
+export type KanoBuildOptions = ElectronBuildOptions & {
+    ['skip-ar']? : boolean;
+}
